+++ driver/web/adapter.go
@@ -108,10 +108,16 @@ func (a Adapter) Start() {
 	adminRouter.HandleFunc("/configs/{id}", a.wrapFunc(a.adminAPIsHandler.updateConfig, a.auth.admin.Permissions)).Methods("PUT")
 	adminRouter.HandleFunc("/configs/{id}", a.wrapFunc(a.adminAPIsHandler.deleteConfig, a.auth.admin.Permissions)).Methods("DELETE")
 
+	//deprecated
 	adminRouter.HandleFunc("/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.addwebtoolsblacklist, a.auth.admin.Permissions)).Methods("PUT")
 	adminRouter.HandleFunc("/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.getwebtoolsblacklist, a.auth.admin.Permissions)).Methods("GET")
 	adminRouter.HandleFunc("/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.removewebtoolsblacklist, a.auth.admin.Permissions)).Methods("DELETE")
-	adminRouter.HandleFunc("/webtools-summary", a.wrapFunc(a.adminAPIsHandler.getWebtoolsSummary, a.auth.admin.Permissions)).Methods("GET")
+
+	adminRouter.HandleFunc("/events/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.addwebtoolsblacklist, a.auth.admin.Permissions)).Methods("PUT")
+	adminRouter.HandleFunc("/events/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.getwebtoolsblacklist, a.auth.admin.Permissions)).Methods("GET")
+	adminRouter.HandleFunc("/events/webtools-blacklist", a.wrapFunc(a.adminAPIsHandler.removewebtoolsblacklist, a.auth.admin.Permissions)).Methods("DELETE")
+	adminRouter.HandleFunc("/events/summary", a.wrapFunc(a.adminAPIsHandler.getEventsSummary, a.auth.admin.Permissions)).Methods("GET")
+	//adminRouter.HandleFunc("/legacy-events", a.wrapFunc(a.adminAPIsHandler.legacyEvents, a.auth.admin.Permissions)).Methods("GET")
 
 	// BB APIs
 	bbsRouter := mainRouter.PathPrefix("/bbs").Subrouter()