+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Expose get summary events API [#135](https://github.com/rokwire/gateway-building-block/issues/135)
+
 ### Changed
 - Remove temporary code for data setup [#137](https://github.com/rokwire/gateway-building-block/issues/137)